refactor(bridge): extract media message conversion
